Heritability in plant breeding

Heritability in plant breeding on a genotype difference basis

H2cal(
  data,
  trait,
  gen.name,
  rep.n,
  loc.n = 1,
  year.n = 1,
  loc.name = NULL,
  year.name = NULL,
  fix.model,
  ran.model,
  summary = FALSE,
  blues = FALSE,
  effects = FALSE,
  plot_diag = FALSE,
  plot_dots = NULL,
  outliers.rm = FALSE
)

Arguments

data
Experimental design data frame with the factors and traits.

trait
Name of the trait.

gen.name
Name of the genotypes.

rep.n
Number of replications in the experiment.

loc.n
Number of locations (default = 1). See details.

year.n
Number of years (default = 1). See details.

loc.name
Name of the location (default = NULL). See details.

year.name
Name of the years (default = NULL). See details.

fix.model
The fixed effects in the model. See examples.

ran.model
The random effects in the model. See examples.

summary
Print summary from random model (default = FALSE).

blues
Calculate the BLUEs (default = FALSE).

effects
Random effects instead of the BLUPs (default = FALSE).

plot_diag
Show diagnostic plots (default = FALSE).

plot_dots
Show dotplot genotypes vs trait (default = NULL). See examples.

outliers.rm
Remove outliers (default = FALSE). See references.

Value

list

Details

The function allows to made the calculation for individual or multi-environmental trials (MET) using th fixed and random model.

1. The variance components.

2. Heritability under three approaches: Standard, Cullis and Piepho.

3. Best Linear Unbiased Predictors (BLUPs).

4. Best Linear Unbiased Estimators (BLUEs).

5. Outliers remove.

For individual experiments is necessary provide the trait, gen.name, rep.n.

For MET experiments you should loc.n and loc.name and/or year.n and year.name according your experiment.

The blues calculation is based in the pairwise comparison and its could takes time according the number of the genotypes.

You can specify as blues = FALSE for calculate the variance components and blups faster.

For more information review the references.

References

Schmidt, P., J. Hartung, J. Bennewitz, and H.P. Piepho. 2019. Heritability in Plant Breeding on a Genotype Difference Basis. Genetics 212(4).

Schmidt, P., J. Hartung, J. Rath, and H.P. Piepho. 2019. Estimating Broad Sense Heritability with Unbalanced Data from Agricultural Cultivar Trials. Crop Science 59(2).

Bernal Vasquez, Angela Maria, et al. “Outlier Detection Methods for Generalized Lattices: A Case Study on the Transition from ANOVA to REML.” Theoretical and Applied Genetics, vol. 129, no. 4, Apr. 2016.

Author

Maria Belen Kistner

Flavio Lozano Isla

Examples

library(inti)
library(agridat)

 dt <- john.alpha
 hr <- H2cal(data = dt
            , trait = "yield"
            , gen.name = "gen"
            , rep.n = 3
            , fix.model = "rep + (1|rep:block) + gen"
            , ran.model = "rep + (1|rep:block) + (1|gen)"
            , blues = TRUE
            , plot_diag = FALSE
            , plot_dots = "rep"
            , outliers.rm = TRUE
            )
 hr$tabsmr

#> # A tibble: 1 x 16
#>   variable   rep  geno   env  year  mean   std   min   max   V.g V.gxl V.gxy
#>   <chr>    <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 yield        3    24     1     1  4.43 0.420  3.44  5.04 0.146     0     0
#> # ... with 4 more variables: V.e <dbl>, h2.s <dbl>, h2.c <dbl>, h2.p <dbl>

 hr$blups

#> # A tibble: 24 x 2
#>    gen   yield
#>    <chr> <dbl>
#>  1 G01    4.97
#>  2 G02    4.47
#>  3 G03    3.62
#>  4 G04    4.58
#>  5 G05    4.57
#>  6 G06    4.50
#>  7 G07    4.10
#>  8 G08    4.57
#>  9 G09    3.56
#> 10 G10    4.33
#> # ... with 14 more rows

 hr$blues

#> # A tibble: 24 x 6
#>    gen   yield    SE    df lower.CL upper.CL
#>    <fct> <dbl> <dbl> <dbl>    <dbl>    <dbl>
#>  1 G01    5.04 0.129  42.6     4.78     5.30
#>  2 G02    4.43 0.129  42.6     4.17     4.69
#>  3 G03    3.53 0.131  42.7     3.27     3.80
#>  4 G04    4.59 0.131  42.7     4.33     4.86
#>  5 G05    4.55 0.154  40.6     4.23     4.86
#>  6 G06    4.55 0.130  42.6     4.29     4.81
#>  7 G07    4.05 0.129  42.6     3.79     4.31
#>  8 G08    4.25 0.153  40.5     3.94     4.56
#>  9 G09    3.44 0.131  42.7     3.17     3.70
#> 10 G10    4.30 0.129  42.6     4.04     4.56
#> # ... with 14 more rows