Field book experimental designs

Function to deploy experimental designs based in agricolae package

fieldbook_design(
  data,
  n_factors = 1,
  type = "crd",
  rep = 2,
  serie = 2,
  seed = 0
)

Arguments

data
Experimental design data frame with the factors and level. See examples.

n_factors
Number of factor in the experiment(default = 1). See details.

type
Type of experimental arrange (default = "crd"). See details.

rep
Number of replications in the experiment (default = 3).

serie
Digits in the plot id (default = 2).

seed
Replicability of draw results (default = 0) always random. See details.

Value

A list with the fieldbook design

Details

The function allows to include the arguments in the sheet that have the information of the design. You should include 2 columns in the sheet: {arguments} and {values}. See examples. The information will be extracted automatically and deploy the design. n_factors = 1: crd, rcbd, lsd, lattice. n_factors = 2 (factorial): split_crd, split_rcbd n_factors >= 2 (factorial): crd, rcbd, lsd.

Examples

# \donttest{

library(inti)
library(googlesheets4)
library(dplyr)

url <- paste0("https://docs.google.com/spreadsheets/d/"
, "1ilw0NHT7mihaM-3U48KzkuMt927xe8ukX6rNuIw2fT0/edit#gid=0")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs %>%
    range_read("tarpuy"))

#> Using an auto-discovered, cached token.
#> To suppress this message, modify your code or options to clearly consent to the use of a cached token.
#> See gargle's "Non-interactive auth" vignette for more details:
#> https://gargle.r-lib.org/articles/non-interactive-auth.html

#> The googlesheets4 package is using a cached token for flavjack@gmail.com.

#> Reading from "inti"

#> Range "'tarpuy'"

#> # A tibble: 16 x 5
#>     temp  prof condición genotype profundidad
#>    <dbl> <dbl> <chr>     <chr>    <chr>      
#>  1    15   0.3 luz       G01      1cm        
#>  2    20   0.6 oscuridad G02      2cm        
#>  3    25   0.7 NA        G03      3cm        
#>  4    30  NA   NA        G04      NA         
#>  5    35  NA   NA        G05      NA         
#>  6    NA  NA   NA        G06      NA         
#>  7    NA  NA   NA        G07      NA         
#>  8    NA  NA   NA        G08      NA         
#>  9    NA  NA   NA        G09      NA         
#> 10    NA  NA   NA        G10      NA         
#> 11    NA  NA   NA        G11      NA         
#> 12    NA  NA   NA        G12      NA         
#> 13    NA  NA   NA        G13      NA         
#> 14    NA  NA   NA        G14      NA         
#> 15    NA  NA   NA        G15      NA         
#> 16    NA  NA   NA        G16      NA         

data %>% inti::fieldbook_design()

#> $design
#>    plots r temp
#> 1    101 1   25
#> 2    102 1   30
#> 3    103 1   15
#> 4    104 2   15
#> 5    105 1   20
#> 6    106 1   35
#> 7    107 2   35
#> 8    108 2   20
#> 9    109 2   25
#> 10   110 2   30
#> 

# }