Fieldbook reshape

Function to reshape fieldbook according a separation character

fieldbook_reshape(
  data,
  last_factor,
  sep,
  new_colname,
  from_var = NULL,
  to_var = NULL,
  exc_factors = NULL
)

Arguments

data
Field book raw data.

last_factor
The last factor in your field book.

sep
Character that separates the last value.

new_colname
The new name for the column created.

from_var
The first variable in case you want to exclude several. variables.

to_var
The last variable in case you want to exclude several variables.

exc_factors
Factor to exclude during the reshape.

Value

data frame

Details

If you variable name is variable_evaluation_rep. The reshape function will help to create the column rep and the new variable name will be variable_evaluation.

Examples

# \donttest{

library(inti)
library(googlesheets4)

url <- paste0("https://docs.google.com/spreadsheets/d/"
              , "1gue-wSQcEu4nJigVZdUWsTfIIzhtxpDRdWAQiEHgKak/edit#gid=56106366")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs %>%
    range_read("fb"))

#> Reading from "RECIFE-2015-06-07 (J. Curcas - Imbibition time)"

#> Range "'fb'"

#> # A tibble: 52 x 43
#>    bar_code plots block imb   imbt  sch_0_1 sch_0_2 sch_0_3 hpt_0_1 elc_0_1
#>    <chr>    <chr> <chr> <chr> <chr>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 PE2015_~ 101   I     8.00  8hrs     59.2    44.2    57.4    7.57  0.448 
#>  2 PE2015_~ 102   I     0.00  0hrs     64.4    88.4    35.8    5.18  0.0218
#>  3 PE2015_~ 103   I     12.00 12hrs    37.6    54.1    63.0    7.08  0.443 
#>  4 PE2015_~ 104   I     18.00 18hrs    56.3    69.4    44.4    6.88  0.662 
#>  5 PE2015_~ 105   I     4.00  4hrs     67.6    38.6    30.7    7.6   0.48  
#>  6 PE2015_~ 106   I     2.00  2hrs     54.8    38.2    50.0    7.91  0.32  
#>  7 PE2015_~ 107   I     6.00  6hrs     56.4    54.8    40.8    7.34  0.461 
#>  8 PE2015_~ 108   I     10.00 10hrs    39.2    55.6    58.0    7.23  0.503 
#>  9 PE2015_~ 109   I     22.00 22hrs    73.8    60.6    56.8    6.89  0.55  
#> 10 PE2015_~ 110   I     20.00 20hrs    75.6    75.8    86.2    6.8   0.61  
#> # ... with 42 more rows, and 33 more variables: tct_0_1 <dbl>, sfw_0_1 <dbl>,
#> #   stw_0_1 <dbl>, sdw_0_1 <dbl>, smt_0_1 <dbl>, swc_0_1 <dbl>,
#> #   nseed_0_1 <dbl>, sgr_0_1 <lgl>, sgr_1_1 <lgl>, sgr_2_1 <dbl>,
#> #   sgr_3_1 <dbl>, sgr_4_1 <dbl>, sgr_5_1 <dbl>, sgr_6_1 <dbl>, sgr_7_1 <dbl>,
#> #   sgr_8_1 <dbl>, sgr_9_1 <dbl>, sgr_10_1 <dbl>, sgr_11_1 <dbl>,
#> #   sgr_12_1 <dbl>, sgr_13_1 <lgl>, sgr_14_1 <lgl>, sgr_15_1 <dbl>,
#> #   sgr_16_1 <dbl>, sgr_17_1 <dbl>, sgr_18_1 <dbl>, sgr_19_1 <lgl>,
#> #   sgr_20_1 <dbl>, sgr_21_1 <lgl>, sgr_22_1 <lgl>, sgr_23_1 <lgl>,
#> #   sgr_24_1 <lgl>, sgr_25_1 <lgl>

fbrs <- fieldbook_reshape(data = data
                          , last_factor = "imbt"
                          , sep = "_"
                          , new_colname = "rep"
                          , from_var = "sch_0_1"
                          , to_var = "swc_0_1"
                          , exc_factors = "bar_code"
                          )
fbrs

#> # A tibble: 156 x 14
#>    plots block imb   imbt  rep   sch_0 hpt_0   elc_0 tct_0 sfw_0 stw_0 sdw_0
#>    <chr> <chr> <chr> <chr> <chr> <dbl> <dbl>   <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 101   I     8.00  8hrs  1      59.2  7.57  0.448      1  5.55  7.61  5.04
#>  2 101   I     8.00  8hrs  2      44.2 NA    NA         NA NA    NA    NA   
#>  3 101   I     8.00  8hrs  3      57.4 NA    NA         NA NA    NA    NA   
#>  4 102   I     0.00  0hrs  1      64.4  5.18  0.0218     1  6.30  6.30  5.80
#>  5 102   I     0.00  0hrs  2      88.4 NA    NA         NA NA    NA    NA   
#>  6 102   I     0.00  0hrs  3      35.8 NA    NA         NA NA    NA    NA   
#>  7 103   I     12.00 12hrs 1      37.6  7.08  0.443      1  6.38  9.13  5.82
#>  8 103   I     12.00 12hrs 2      54.1 NA    NA         NA NA    NA    NA   
#>  9 103   I     12.00 12hrs 3      63.0 NA    NA         NA NA    NA    NA   
#> 10 104   I     18.00 18hrs 1      56.3  6.88  0.662      1  6.63 10.2   6.04
#> # ... with 146 more rows, and 2 more variables: smt_0 <dbl>, swc_0 <dbl>

# }