Fieldbook variable list

Function to include the variables to evaluate in the fieldbook design.

fieldbook_varlist(fieldbook, varlist = NULL)

Arguments

fieldbook
Data frame with the fieldbook.

varlist
Data frame with the variables information. See examples.

Value

data frame

Details

The function allows to include the arguments in the sheet that have the information of the variables. You should include 3 columns in the sheet: {abbreviation}, {evaluation} and {sampling}. See examples. The information will be extracted automatically and deploy the list of variable for the fieldbook design. If you will be evaluate the experimental unit one time you should set 1.

Examples

# \donttest{

library(inti)
library(googlesheets4)

url <- paste0("https://docs.google.com/spreadsheets/d/"
, "1a82XIbTeWPC4pwvu9Zjl4qrGitGQ_B-mf3-w67VS4-Q/edit#gid=0")
# browseURL(url)
gs <- as_sheets_id(url)

(design <- gs %>%
    range_read("design"))

#> Reading from "TARPUY"

#> Range "'design'"

#> # A tibble: 5 x 4
#>   `{arguments}` `{values}`  temp cond    
#>   <chr>         <chr>      <dbl> <chr>   
#> 1 nfactors      2             20 luz     
#> 2 type          rcbd          25 ocuridad
#> 3 rep           4             30 NA      
#> 4 serie         2             35 NA      
#> 5 seed          0             NA NA      

(varlist <- gs %>%
    range_read("varlist"))

#> Reading from "TARPUY"

#> Range "'varlist'"

#> # A tibble: 10 x 6
#>    variable      `{siglas}` `{evaluation}` `{sampling}` units description       
#>    <chr>         <chr>      <chr>                 <dbl> <chr> <chr>             
#>  1 floración     flw        NA                       NA NA    Estado fenológico 
#>  2 madurez       mtr        NA                       NA NA    Estado fenológico 
#>  3 cosecha       hrv        NA                       NA NA    Estado fenológico 
#>  4 días a la fl~ dflw       flw                       1 días  Días después de l~
#>  5 altura de pl~ altp       flw                       3 cm    Desde la base has~
#>  6 diámetro del~ dmt        flw                       3 mm    Usar un vernier   
#>  7 días a la ma~ dmdf       mtr                       1 días  Días después de l~
#>  8 altura de pl~ altp       mtr                       3 cm    Desde la base has~
#>  9 diámetro del~ dmt        mtr                       3 mm    Días después de l~
#> 10 rendimiento ~ r10p       hrv                       2 g     Seleccionar las p~

fieldbook <- design %>%
  inti::fieldbook_design() %>%
  inti::fieldbook_varlist(varlist)

fieldbook

#> $design
#>    plots block temp     cond dflw_flw_1 altp_flw_1 altp_flw_2 altp_flw_3
#> 1    101     1   20      luz         NA         NA         NA         NA
#> 2    102     1   25      luz         NA         NA         NA         NA
#> 3    103     1   20 ocuridad         NA         NA         NA         NA
#> 4    104     1   30      luz         NA         NA         NA         NA
#> 5    105     1   30 ocuridad         NA         NA         NA         NA
#> 6    106     1   35 ocuridad         NA         NA         NA         NA
#> 7    107     1   35      luz         NA         NA         NA         NA
#> 8    108     1   25 ocuridad         NA         NA         NA         NA
#> 9    109     2   25 ocuridad         NA         NA         NA         NA
#> 10   110     2   25      luz         NA         NA         NA         NA
#> 11   111     2   30 ocuridad         NA         NA         NA         NA
#> 12   112     2   30      luz         NA         NA         NA         NA
#> 13   113     2   20      luz         NA         NA         NA         NA
#> 14   114     2   20 ocuridad         NA         NA         NA         NA
#> 15   115     2   35      luz         NA         NA         NA         NA
#> 16   116     2   35 ocuridad         NA         NA         NA         NA
#> 17   117     3   30      luz         NA         NA         NA         NA
#> 18   118     3   25 ocuridad         NA         NA         NA         NA
#> 19   119     3   20 ocuridad         NA         NA         NA         NA
#> 20   120     3   25      luz         NA         NA         NA         NA
#> 21   121     3   35 ocuridad         NA         NA         NA         NA
#> 22   122     3   30 ocuridad         NA         NA         NA         NA
#> 23   123     3   35      luz         NA         NA         NA         NA
#> 24   124     3   20      luz         NA         NA         NA         NA
#> 25   125     4   30 ocuridad         NA         NA         NA         NA
#> 26   126     4   20 ocuridad         NA         NA         NA         NA
#> 27   127     4   25 ocuridad         NA         NA         NA         NA
#> 28   128     4   25      luz         NA         NA         NA         NA
#> 29   129     4   20      luz         NA         NA         NA         NA
#> 30   130     4   35 ocuridad         NA         NA         NA         NA
#> 31   131     4   30      luz         NA         NA         NA         NA
#> 32   132     4   35      luz         NA         NA         NA         NA
#>    dmt_flw_1 dmt_flw_2 dmt_flw_3 dmdf_mtr_1 altp_mtr_1 altp_mtr_2 altp_mtr_3
#> 1         NA        NA        NA         NA         NA         NA         NA
#> 2         NA        NA        NA         NA         NA         NA         NA
#> 3         NA        NA        NA         NA         NA         NA         NA
#> 4         NA        NA        NA         NA         NA         NA         NA
#> 5         NA        NA        NA         NA         NA         NA         NA
#> 6         NA        NA        NA         NA         NA         NA         NA
#> 7         NA        NA        NA         NA         NA         NA         NA
#> 8         NA        NA        NA         NA         NA         NA         NA
#> 9         NA        NA        NA         NA         NA         NA         NA
#> 10        NA        NA        NA         NA         NA         NA         NA
#> 11        NA        NA        NA         NA         NA         NA         NA
#> 12        NA        NA        NA         NA         NA         NA         NA
#> 13        NA        NA        NA         NA         NA         NA         NA
#> 14        NA        NA        NA         NA         NA         NA         NA
#> 15        NA        NA        NA         NA         NA         NA         NA
#> 16        NA        NA        NA         NA         NA         NA         NA
#> 17        NA        NA        NA         NA         NA         NA         NA
#> 18        NA        NA        NA         NA         NA         NA         NA
#> 19        NA        NA        NA         NA         NA         NA         NA
#> 20        NA        NA        NA         NA         NA         NA         NA
#> 21        NA        NA        NA         NA         NA         NA         NA
#> 22        NA        NA        NA         NA         NA         NA         NA
#> 23        NA        NA        NA         NA         NA         NA         NA
#> 24        NA        NA        NA         NA         NA         NA         NA
#> 25        NA        NA        NA         NA         NA         NA         NA
#> 26        NA        NA        NA         NA         NA         NA         NA
#> 27        NA        NA        NA         NA         NA         NA         NA
#> 28        NA        NA        NA         NA         NA         NA         NA
#> 29        NA        NA        NA         NA         NA         NA         NA
#> 30        NA        NA        NA         NA         NA         NA         NA
#> 31        NA        NA        NA         NA         NA         NA         NA
#> 32        NA        NA        NA         NA         NA         NA         NA
#>    dmt_mtr_1 dmt_mtr_2 dmt_mtr_3 r10p_hrv_1 r10p_hrv_2
#> 1         NA        NA        NA         NA         NA
#> 2         NA        NA        NA         NA         NA
#> 3         NA        NA        NA         NA         NA
#> 4         NA        NA        NA         NA         NA
#> 5         NA        NA        NA         NA         NA
#> 6         NA        NA        NA         NA         NA
#> 7         NA        NA        NA         NA         NA
#> 8         NA        NA        NA         NA         NA
#> 9         NA        NA        NA         NA         NA
#> 10        NA        NA        NA         NA         NA
#> 11        NA        NA        NA         NA         NA
#> 12        NA        NA        NA         NA         NA
#> 13        NA        NA        NA         NA         NA
#> 14        NA        NA        NA         NA         NA
#> 15        NA        NA        NA         NA         NA
#> 16        NA        NA        NA         NA         NA
#> 17        NA        NA        NA         NA         NA
#> 18        NA        NA        NA         NA         NA
#> 19        NA        NA        NA         NA         NA
#> 20        NA        NA        NA         NA         NA
#> 21        NA        NA        NA         NA         NA
#> 22        NA        NA        NA         NA         NA
#> 23        NA        NA        NA         NA         NA
#> 24        NA        NA        NA         NA         NA
#> 25        NA        NA        NA         NA         NA
#> 26        NA        NA        NA         NA         NA
#> 27        NA        NA        NA         NA         NA
#> 28        NA        NA        NA         NA         NA
#> 29        NA        NA        NA         NA         NA
#> 30        NA        NA        NA         NA         NA
#> 31        NA        NA        NA         NA         NA
#> 32        NA        NA        NA         NA         NA
#> 

# }