Figure with caption and notes

Include figures with title and notes using a data base

include_fig(
  data = NULL,
  figure,
  caption = NULL,
  notes = NULL,
  label = "Source:",
  table = c("figures", "caption", "notes", "url")
)

Arguments

data
Data frame with the figures information. See details.

figure
Name or path of the figure.

caption
Manual figure caption (default = NULL).

notes
Manual figure notes (default = NULL).

label
Label for start the footnote (default = "Source:").

table
Columns in the table (default = c("figures", "caption", "notes", "url")).

Value

Figure with caption and notes

Details

The data frame should contain 4 columns specified in table:

1. figures.

2. caption.

3. notes.

4. url or url.

If you don't use a a data frame you provide the information manually.