Remove outliers

Use the method M4 in Bernal Vasquez (2016). Bonferroni Holm test to judge residuals standardized by the re scaled MAD (BH MADR).

outliers_remove(data, trait, model)

Arguments

data
Experimental design data frame with the factors and traits.

trait
Name of the trait.

model
The fixed or random effects in the model.

Value

list. 1. Table with date without outliers. 2. The outliers in the dataset.

Details

Function to remove outliers in MET experiments

References

Bernal Vasquez, Angela Maria, et al. “Outlier Detection Methods for Generalized Lattices: A Case Study on the Transition from ANOVA to REML.” Theoretical and Applied Genetics, vol. 129, no. 4, Apr. 2016.

Examples

library(inti)
library(agridat)

outliers <- outliers_remove(
  data = john.alpha
  , trait ="yield"
  , model = "rep + (1|rep:block) + gen"
  )
  
outliers

#> $data
#>    rep block gen  yield
#> 1   R1    B1 G11 4.1172
#> 2   R1    B1 G04 4.4461
#> 4   R1    B1 G22 4.5784
#> 5   R1    B2 G21 4.6540
#> 6   R1    B2 G10 4.1736
#> 7   R1    B2 G20 4.0141
#> 8   R1    B2 G02 4.3350
#> 9   R1    B3 G23 4.2323
#> 10  R1    B3 G14 4.7572
#> 11  R1    B3 G16 4.4906
#> 12  R1    B3 G18 3.9737
#> 13  R1    B4 G13 4.2530
#> 14  R1    B4 G03 3.3420
#> 15  R1    B4 G19 4.7269
#> 17  R1    B5 G17 4.7876
#> 18  R1    B5 G15 5.0902
#> 19  R1    B5 G07 4.1505
#> 20  R1    B5 G01 5.1202
#> 21  R1    B6 G06 4.7085
#> 22  R1    B6 G12 5.2560
#> 24  R1    B6 G09 3.3986
#> 25  R2    B1 G08 3.9926
#> 26  R2    B1 G20 3.6056
#> 27  R2    B1 G14 4.5294
#> 28  R2    B1 G04 4.3599
#> 29  R2    B2 G24 3.9039
#> 30  R2    B2 G15 4.9114
#> 31  R2    B2 G03 3.7999
#> 32  R2    B2 G23 4.3042
#> 33  R2    B3 G12 5.3127
#> 34  R2    B3 G11 5.1163
#> 35  R2    B3 G21 5.3802
#> 36  R2    B3 G17 5.0744
#> 37  R2    B4 G05 5.1202
#> 38  R2    B4 G09 4.2955
#> 39  R2    B4 G10 4.9057
#> 40  R2    B4 G01 5.7161
#> 41  R2    B5 G02 5.1566
#> 42  R2    B5 G18 5.0988
#> 43  R2    B5 G13 5.4840
#> 44  R2    B5 G22 5.0969
#> 45  R2    B6 G19 5.3148
#> 46  R2    B6 G07 4.6297
#> 47  R2    B6 G06 5.1751
#> 48  R2    B6 G16 5.3024
#> 49  R3    B1 G11 3.9205
#> 50  R3    B1 G01 4.6512
#> 51  R3    B1 G14 4.3887
#> 52  R3    B1 G19 4.5552
#> 53  R3    B2 G02 4.0510
#> 54  R3    B2 G15 4.6783
#> 55  R3    B2 G09 3.1407
#> 56  R3    B2 G08 3.9821
#> 57  R3    B3 G17 4.3234
#> 58  R3    B3 G18 4.2486
#> 59  R3    B3 G04 4.3960
#> 60  R3    B3 G06 4.2474
#> 61  R3    B4 G12 4.1746
#> 62  R3    B4 G13 4.7512
#> 63  R3    B4 G10 4.0875
#> 64  R3    B4 G23 3.8721
#> 65  R3    B5 G21 4.4130
#> 66  R3    B5 G22 4.2397
#> 67  R3    B5 G16 4.3852
#> 68  R3    B5 G24 3.5655
#> 69  R3    B6 G03 2.8873
#> 70  R3    B6 G05 4.1972
#> 71  R3    B6 G20 3.7349
#> 72  R3    B6 G07 3.6096
#> 
#> $outliers
#>    rep block gen  yield      resi  res_MAD  rawp.BHStud index         adjp
#> 3   R1    B1 G05 5.8757 0.7022176 4.929351 8.250342e-07     3 8.250342e-07
#> 16  R1    B4 G08 4.9989 0.5179535 3.635874 2.770401e-04    16 2.770401e-04
#> 23  R1    B6 G24 4.9577 0.5385472 3.780435 1.565546e-04    23 1.565546e-04
#>           bholm out_flag
#> 3  5.940247e-05  OUTLIER
#> 16 1.939281e-02  OUTLIER
#> 23 1.111538e-02  OUTLIER
#>