Heritability in plant breeding

Heritability in plant breeding on a genotype difference basis

H2cal(
  data,
  trait,
  gen.name,
  rep.n,
  loc.n = 1,
  year.n = 1,
  loc.name = NULL,
  year.name = NULL,
  fix.model,
  ran.model,
  blues = FALSE,
  effects = FALSE,
  plot_diag = FALSE,
  plot_dots = NULL,
  outliers.rm = FALSE
)

Arguments

data
Experimental design data frame with the factors and traits.

trait
Name of the trait.

gen.name
Name of the genotypes.

rep.n
Number of replications in the experiment.

loc.n
Number of locations (default = 1). See details.

year.n
Number of years (default = 1). See details.

loc.name
Name of the location (default = NULL). See details.

year.name
Name of the years (default = NULL). See details.

fix.model
The fixed effects in the model. See examples.

ran.model
The random effects in the model. See examples.

blues
Calculate the BLUEs (default = FALSE).

effects
Random effects instead of the BLUPs (default = FALSE).

plot_diag
Show diagnostic plots (default = FALSE).

plot_dots
Show dotplot genotypes vs trait (default = NULL). See examples.

outliers.rm
Remove outliers (default = FALSE). See references.

Value

list

Details

The function allows to made the calculation for individual or multi-environmental trials (MET) using th fixed and random model.

1. The variance components.

2. Heritability under three approaches: Standard, Cullis and Piepho.

3. Best Linear Unbiased Predictors (BLUPs).

4. Best Linear Unbiased Estimators (BLUEs).

5. Outliers remove.

For individual experiments is necessary provide the trait, gen.name, rep.n.

For MET experiments you should loc.n and loc.name and/or year.n and year.name according your experiment.

The blues calculation is based in the pairwise comparison and its could takes time according the number of the genotypes.

You can specify as blues = FALSE for calculate the variance components and blups faster.

For more information review the references.

References

Schmidt, P., J. Hartung, J. Bennewitz, and H.P. Piepho. 2019. Heritability in Plant Breeding on a Genotype Difference Basis. Genetics 212(4).

Schmidt, P., J. Hartung, J. Rath, and H.P. Piepho. 2019. Estimating Broad Sense Heritability with Unbalanced Data from Agricultural Cultivar Trials. Crop Science 59(2).

Bernal Vasquez, Angela Maria, et al. “Outlier Detection Methods for Generalized Lattices: A Case Study on the Transition from ANOVA to REML.” Theoretical and Applied Genetics, vol. 129, no. 4, Apr. 2016.

Author

Maria Belen Kistner

Flavio Lozano Isla

Examples

if (FALSE) {
library(inti)
library(agridat)

 dt <- john.alpha
 hr <- H2cal(data = dt
            , trait = "yield"
            , gen.name = "gen"
            , rep.n = 3
            , fix.model = "rep + (1|rep:block) + gen"
            , ran.model = "rep + (1|rep:block) + (1|gen)"
            , blues = T
            , plot_diag = TRUE
            , plot_dots = "rep"
            , outliers.rm = T
            )
 hr$tabsmr
 hr$blups
 hr$blues
}