Field book experimental designs

Function to deploy experimental designs based in agricolae package

fieldbook_design(
  data,
  n_factors = 1,
  type = "crd",
  rep = 2,
  serie = 2,
  seed = 0
)

Arguments

data
Experimental design data frame with the factors and level. See examples.

n_factors
Number of factor in the experiment(default = 1). See details.

type
Type of experimental arrange (default = "crd"). See details.

rep
Number of replications in the experiment (default = 3).

serie
Digits in the plot id (default = 2).

seed
Replicability of draw results (default = 0) always random. See details.

Value

A list with the fieldbook design

Details

The function allows to include the arguments in the sheet that have the information of the design. You should include 2 columns in the sheet: {arguments} and {values}. See examples. The information will be extracted automatically and deploy the design. n_factors = 1: crd, rcbd, lsd, lattice. n_factors = 2 (factorial): split_crd, split_rcbd n_factors >= 2 (factorial): crd, rcbd, lsd.

Examples

# \donttest{

library(inti)
library(googlesheets4)
library(dplyr)

url <- paste0("https://docs.google.com/spreadsheets/d/"
, "1ilw0NHT7mihaM-3U48KzkuMt927xe8ukX6rNuIw2fT0/edit#gid=0")
# browseURL(url)
gs <- as_sheets_id(url)

(data <- gs %>%
    range_read("tarpuy"))

#> Using an auto-discovered, cached token.
#> To suppress this message, modify your code or options to clearly consent to the use of a cached token.
#> See gargle's "Non-interactive auth" vignette for more details:
#> https://gargle.r-lib.org/articles/non-interactive-auth.html

#> The googlesheets4 package is using a cached token for flavjack@gmail.com.

#> Error: Client error: (429) RESOURCE_EXHAUSTED
#>   * Either out of resource quota or reaching rate limiting. The client should look for google.rpc.QuotaFailure error detail for more information.
#>   * Quota exceeded for quota group 'ReadGroup' and limit 'Read requests per 100 seconds' of service 'sheets.googleapis.com' for consumer 'project_number:603366585132'.
#> 
#> Error details:
#>   * Error details of type 'google.rpc.Help' are not implemented yet.
#>   * Workaround: use `tryCatch()` and inspect error payload yourself.
#>   * Please open an issue at https://github.com/r-lib/gargle/issues, so we can fix.

data %>% inti::fieldbook_design()

#> Warning: `select_()` is deprecated as of dplyr 0.7.0.
#> Please use `select()` instead.
#> This warning is displayed once every 8 hours.
#> Call `lifecycle::last_warnings()` to see where this warning was generated.

#> Error in UseMethod("select_"): no applicable method for 'select_' applied to an object of class "function"

# }